requirements.txt
click==8.1.8
colorama==0.4.6
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.2
pillow==9.5.0
qrcode==7.4.2
pymysql
flask-httpauth
python-dotenv
pytz==2023.3

